+++ app/src/main/java/org/inspire/breath/fragments/home/Listings.java
@@ -1,9 +1,8 @@
-package org.inspire.breath.fragments.patients;
+package org.inspire.breath.fragments.home;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -11,14 +10,17 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import org.inspire.breath.R;
-import org.inspire.breath.activities.PatientsActivity;
 import org.inspire.breath.adapters.PatientListAdapter;
 import org.inspire.breath.data.AppRoomDatabase;
 import org.inspire.breath.data.Patient;
+import org.inspire.breath.views.StaticPager;
+
+import android.support.v4.app.Fragment;
+
 
 import java.util.List;
 
-public class Listings extends PatientsFragment implements PatientListAdapter.PatientCallback {
+public class Listings extends Fragment implements PatientListAdapter.PatientCallback, StaticPager.Focusable{
 
     private RecyclerView mPatientList;
 
@@ -59,7 +61,6 @@ public class Listings extends PatientsFragment implements PatientListAdapter.Pat
 
     @Override
     public void onFocus() {
-        super.onFocus();
         updateRecycler();
     }
