--- app/src/main/java/org/inspire/breath/fragments/patients/PatientsFragment.java
@@ -1,17 +1,0 @@
-package org.inspire.breath.fragments.patients;
-
-import android.support.v4.app.Fragment;
-
-import org.inspire.breath.activities.PatientsActivity;
-import org.inspire.breath.views.StaticPager;
-
-public class PatientsFragment extends Fragment implements StaticPager.Focusable {
-
-    PatientsActivity getPatientsActivity() {
-        return (PatientsActivity) getActivity();
-    }
-
-    @Override
-    public void onFocus() {
-    }
-}
